Fix new trade page dynamic hydration
