+++ app/new-trade/page.tsx
@@ -1,392 +1,7 @@
-"use client";
-
 export const dynamic = "force-dynamic";
 
-import { useRouter, useSearchParams } from "next/navigation";
-import { useEffect, useMemo, useState } from "react";
-
-import Button from "@/components/ui/Button";
-import {
-  appendStoredTrade,
-  findStoredTrade,
-  type SymbolOption,
-  type StoredTrade,
-  updateStoredTrade,
-} from "@/lib/tradesStorage";
-import { SYMBOL_OPTIONS } from "@/lib/symbols";
-
-const availableSymbols: SymbolOption[] = SYMBOL_OPTIONS;
-
-function getWeekDays(baseDate: Date) {
-  const reference = new Date(baseDate);
-  reference.setHours(0, 0, 0, 0);
-  const diffFromMonday = (reference.getDay() + 6) % 7;
-  const monday = new Date(reference);
-  monday.setDate(reference.getDate() - diffFromMonday);
-
-  return Array.from({ length: 5 }, (_, index) => {
-    const date = new Date(monday);
-    date.setDate(monday.getDate() + index);
-    return date;
-  });
-}
+import NewTradeClient from "./NewTradeClient";
 
 export default function NewTradePage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const today = useMemo(() => {
-    const now = new Date();
-    now.setHours(0, 0, 0, 0);
-    return now;
-  }, []);
-
-  const [weekReference, setWeekReference] = useState<Date>(() => today);
-  const currentWeekDays = useMemo(() => getWeekDays(weekReference), [weekReference]);
-  const [selectedDate, setSelectedDate] = useState(() => {
-    const initialWeek = getWeekDays(today);
-    const dayIndex = Math.min((today.getDay() + 6) % 7, 4);
-    const initialDate = initialWeek.at(dayIndex) ?? initialWeek[0] ?? today;
-    return new Date(initialDate);
-  });
-
-  const [selectedSymbol, setSelectedSymbol] = useState<SymbolOption>(availableSymbols[2]);
-  const [isSymbolListOpen, setIsSymbolListOpen] = useState(true);
-  const [isSaving, setIsSaving] = useState(false);
-  const [existingTrade, setExistingTrade] = useState<StoredTrade | null>(null);
-  const [isEditing, setIsEditing] = useState(() => !searchParams.get("tradeId"));
-
-  const tradeId = searchParams.get("tradeId");
-
-  useEffect(() => {
-    if (!tradeId) {
-      setExistingTrade(null);
-      setIsEditing(true);
-      setWeekReference(today);
-      return;
-    }
-
-    const storedTrade = findStoredTrade(tradeId);
-    if (storedTrade) {
-      const storedDate = new Date(storedTrade.date);
-      if (!Number.isNaN(storedDate.getTime())) {
-        storedDate.setHours(0, 0, 0, 0);
-        setWeekReference(storedDate);
-        setSelectedDate(new Date(storedDate));
-      }
-      setSelectedSymbol(storedTrade.symbol);
-      setExistingTrade(storedTrade);
-      setIsEditing(false);
-    } else {
-      setExistingTrade(null);
-      setIsEditing(true);
-      setWeekReference(today);
-    }
-  }, [tradeId, today]);
-
-  const dayOfWeekLabel = useMemo(
-    () =>
-      selectedDate.toLocaleDateString(undefined, {
-        weekday: "long",
-      }),
-    [selectedDate]
-  );
-
-  const handleSelectSymbol = (symbol: SymbolOption) => {
-    setSelectedSymbol(symbol);
-  };
-
-  const handleSaveTrade = () => {
-    if (isSaving) return;
-    setIsSaving(true);
-
-    const normalizedDate = new Date(selectedDate);
-    normalizedDate.setHours(0, 0, 0, 0);
-
-    if (existingTrade) {
-      updateStoredTrade(existingTrade.id, {
-        date: normalizedDate.toISOString(),
-        symbol: selectedSymbol,
-      });
-    } else {
-      const trade: StoredTrade = {
-        id:
-          typeof crypto !== "undefined" && "randomUUID" in crypto
-            ? crypto.randomUUID()
-            : Date.now().toString(),
-        date: normalizedDate.toISOString(),
-        symbol: selectedSymbol,
-      };
-
-      appendStoredTrade(trade);
-    }
-
-    router.push("/");
-    setIsSaving(false);
-  };
-
-  return (
-    <section className="relative flex min-h-dvh flex-col overflow-hidden bg-[radial-gradient(circle_at_top,_#ffffff,_#f1f1f1)] px-6 py-10 text-fg">
-      <div className="absolute right-6 top-6 flex flex-col items-end gap-3">
-        {existingTrade ? (
-          <button
-            type="button"
-            onClick={() => {
-              setIsEditing((prev) => {
-                if (prev && existingTrade) {
-                  const storedDate = new Date(existingTrade.date);
-                  if (!Number.isNaN(storedDate.getTime())) {
-                    storedDate.setHours(0, 0, 0, 0);
-                    setWeekReference(storedDate);
-                    setSelectedDate(new Date(storedDate));
-                  }
-                  setSelectedSymbol(existingTrade.symbol);
-                }
-                return !prev;
-              });
-            }}
-            className="flex items-center gap-2 rounded-full border border-border/70 bg-white/80 px-4 py-2 text-sm font-semibold text-muted-fg shadow-sm transition hover:border-border hover:text-fg"
-          >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="1.5"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              className="h-4 w-4"
-              aria-hidden="true"
-            >
-              <path d="M17 3a2.828 2.828 0 1 1 4 4L7.5 20.5 3 21l.5-4.5Z" />
-            </svg>
-            {isEditing ? "Annulla" : "Modifica"}
-          </button>
-        ) : null}
-        <button
-          type="button"
-          className="flex h-12 w-12 items-center justify-center rounded-full border border-border/60 bg-white/70 text-lg font-semibold text-muted-fg shadow-sm transition hover:scale-105 hover:text-fg"
-          onClick={() => {
-            router.back();
-          }}
-          aria-label="Close"
-        >
-          ×
-        </button>
-      </div>
-
-      <div className="mx-auto flex w-full max-w-3xl flex-1 flex-col items-center justify-center gap-12 text-center">
-        <header className="space-y-3">
-          <p className="text-xs font-semibold uppercase tracking-[0.3em] text-muted-fg">Trading Journal</p>
-          <h1 className="text-4xl font-black tracking-tight text-fg drop-shadow-sm md:text-5xl">
-            Register a Trade
-          </h1>
-        </header>
-
-        <div className="flex flex-col items-center gap-8">
-          <nav className="flex items-center gap-3 rounded-full border border-border/60 bg-white/60 px-2 py-2 shadow-lg shadow-black/5 backdrop-blur">
-            {[
-              { label: "Main data", isActive: true },
-              { label: "Performance", isActive: false },
-              { label: "Mindset", isActive: false },
-            ].map(({ label, isActive }) => (
-              <button
-                key={label}
-                type="button"
-                className={`rounded-full px-4 py-2 text-sm font-semibold transition ${
-                  isActive
-                    ? "bg-accent text-white shadow"
-                    : "text-muted-fg hover:text-fg"
-                }`}
-                aria-pressed={isActive}
-                disabled={!isActive}
-              >
-                {label}
-              </button>
-            ))}
-          </nav>
-
-          <div className="w-full rounded-[2.5rem] border border-border/60 bg-white/80 px-4 py-8 shadow-lg shadow-black/10 backdrop-blur">
-            <div className="mx-auto flex max-w-xl items-center gap-2 overflow-x-auto rounded-full bg-transparent px-2 py-1">
-              {currentWeekDays.map((date) => {
-                const isSelected = date.toDateString() === selectedDate.toDateString();
-                const dayNumber = date.getDate();
-                const monthLabel = date
-                  .toLocaleDateString(undefined, {
-                    month: "short",
-                  })
-                  .toUpperCase();
-
-                return (
-                  <button
-                    key={date.toISOString()}
-                    type="button"
-                    onClick={() => setSelectedDate(new Date(date))}
-                    disabled={Boolean(existingTrade) && !isEditing}
-                    className={`flex min-w-[66px] flex-col items-center gap-1 rounded-full px-3 py-2 text-xs font-semibold transition md:min-w-[88px] md:text-sm ${
-                      isSelected
-                        ? "bg-accent text-white shadow"
-                        : "text-muted-fg hover:text-fg"
-                    }`}
-                  >
-                    <span className={`text-xl md:text-2xl ${isSelected ? "font-black" : "font-bold"}`}>
-                      {dayNumber}
-                    </span>
-                    <span className="text-[10px] tracking-[0.3em] md:text-xs">
-                      {monthLabel}
-                    </span>
-                  </button>
-                );
-              })}
-
-              <button
-                type="button"
-                className="ml-auto flex h-14 w-14 flex-none items-center justify-center rounded-full border border-border/70 bg-white text-muted-fg shadow-sm transition hover:text-fg"
-                onClick={() => {
-                  const newReference = new Date(today);
-                  setWeekReference(newReference);
-                  const freshWeek = getWeekDays(newReference);
-                  const dayIndex = Math.min((today.getDay() + 6) % 7, 4);
-                  const targetDate = freshWeek.at(dayIndex) ?? freshWeek[0] ?? newReference;
-                  setSelectedDate(new Date(targetDate));
-                }}
-                aria-label="Select today"
-                title="Select today"
-                disabled={Boolean(existingTrade) && !isEditing}
-              >
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  viewBox="0 0 24 24"
-                  fill="none"
-                  stroke="currentColor"
-                  strokeWidth="1.5"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  className="h-6 w-6"
-                >
-                  <rect x="3" y="4" width="18" height="18" rx="2" ry="2" />
-                  <line x1="16" y1="2" x2="16" y2="6" />
-                  <line x1="8" y1="2" x2="8" y2="6" />
-                  <line x1="3" y1="10" x2="21" y2="10" />
-                  <circle cx="12" cy="16" r="1.5" />
-                </svg>
-              </button>
-            </div>
-
-            <p className="mt-6 text-sm font-medium text-muted-fg md:text-base">
-              Day of the week: <span className="font-semibold capitalize text-fg">{dayOfWeekLabel}</span>
-            </p>
-          </div>
-
-          <div className="w-full rounded-[2.5rem] border border-border/60 bg-white/80 px-6 py-8 shadow-lg shadow-black/10 backdrop-blur">
-            <div className="flex flex-col gap-6">
-              <div className="flex flex-wrap items-center justify-between gap-4">
-                <div className="flex flex-col gap-3">
-                  <span className="text-xs font-semibold uppercase tracking-[0.3em] text-muted-fg">Symbol</span>
-                  <div className="flex items-center gap-3 rounded-2xl border border-border/60 bg-white px-4 py-3 shadow-sm">
-                    <span className="text-2xl" aria-hidden="true">
-                      {selectedSymbol.flag}
-                    </span>
-                    <span className="text-lg font-semibold tracking-[0.2em] text-fg md:text-xl">
-                      {selectedSymbol.code}
-                    </span>
-                  </div>
-                </div>
-                <button
-                  type="button"
-                  onClick={() => setIsSymbolListOpen((prev) => !prev)}
-                  className="ml-auto flex items-center gap-2 rounded-full border border-border/70 bg-white px-4 py-2 text-sm font-semibold text-muted-fg transition hover:border-border hover:text-fg"
-                  aria-haspopup="listbox"
-                  aria-expanded={isSymbolListOpen}
-                  disabled={Boolean(existingTrade) && !isEditing}
-                >
-                  {isSymbolListOpen ? "Hide symbols" : "Show symbols"}
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    stroke="currentColor"
-                    strokeWidth="1.5"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    className={`h-4 w-4 transition-transform ${isSymbolListOpen ? "rotate-180" : ""}`}
-                    aria-hidden="true"
-                  >
-                    <path d="m6 9 6 6 6-6" />
-                  </svg>
-                </button>
-              </div>
-
-              <div
-                className={`grid transition-[grid-template-rows] duration-300 ease-out ${
-                  isSymbolListOpen ? "grid-rows-[1fr]" : "grid-rows-[0fr]"
-                }`}
-              >
-                <div className="overflow-hidden">
-                  <div
-                    className="flex flex-col gap-2 rounded-[1.75rem] border border-border/40 bg-white/80 p-2 shadow-inner shadow-black/5"
-                    role="listbox"
-                    aria-activedescendant={`symbol-${selectedSymbol.code}`}
-                  >
-                    {availableSymbols.map((symbol) => {
-                      const isActive = symbol.code === selectedSymbol.code;
-
-                      return (
-                        <button
-                          key={symbol.code}
-                          id={`symbol-${symbol.code}`}
-                          type="button"
-                          className={`flex w-full items-center gap-4 rounded-2xl px-5 py-3 text-sm font-semibold transition md:text-base ${
-                            isActive
-                              ? "bg-accent text-white shadow-lg shadow-accent/30"
-                              : "bg-transparent text-fg hover:bg-muted/40"
-                          }`}
-                        onClick={() => handleSelectSymbol(symbol)}
-                        disabled={Boolean(existingTrade) && !isEditing}
-                        aria-selected={isActive}
-                        role="option"
-                      >
-                          <span className="text-2xl" aria-hidden="true">
-                            {symbol.flag}
-                          </span>
-                          <span className="tracking-[0.2em]">{symbol.code}</span>
-                          {isActive ? (
-                            <span className="ml-auto text-xs font-bold uppercase tracking-[0.3em] text-white/80">
-                              Selected
-                            </span>
-                          ) : null}
-                        </button>
-                      );
-                    })}
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-
-          {isEditing ? (
-            <div className="w-full rounded-[2.5rem] border border-border/60 bg-white/80 px-6 py-6 shadow-lg shadow-black/10 backdrop-blur">
-              <div className="flex flex-col items-center gap-4 text-center">
-                <h2 className="text-xl font-semibold text-fg">
-                  {existingTrade ? "Ready to update this trade?" : "Ready to save this trade?"}
-                </h2>
-                <p className="text-sm text-muted-fg">
-                  We’ll {existingTrade ? "refresh" : "add"} the selected symbol and date in your journal overview on the home page.
-                </p>
-                <Button
-                  type="button"
-                  variant="primary"
-                  size="lg"
-                  className="rounded-full px-8"
-                  onClick={handleSaveTrade}
-                  disabled={isSaving}
-                >
-                  {isSaving ? "Saving..." : existingTrade ? "Save changes" : "Save"}
-                </Button>
-              </div>
-            </div>
-          ) : null}
-        </div>
-      </div>
-    </section>
-  );
+  return <NewTradeClient />;
 }
